+++ src/pages/TodosPage.jsx
@@ -1,8 +1,8 @@
-import React, {useEffect} from "react";
+import React from "react";
 import TodoList from '../features/TodoList/TodoList.jsx'
 import TodoForm from '../features/TodoForm.jsx'
 import TodosViewForm  from '../features/TodosViewForm.jsx';
-import { useSearchParams, useNavigate } from 'react-router';
+
 function TodosPage({
     todoState,
     addTodo,
@@ -15,39 +15,10 @@ function TodosPage({
     queryString,
     setQueryString,
 }) {
-     const [searchParams, setSearchParams] = useSearchParams();
-     const navigate = useNavigate();
-
-     const itemsPerPage = 15;
-     const currentPage = parseInt(searchParams.get("page") || "1", 10);
-     const filteredTodoList = todoState.todoList.filter((todo) => 
+     
+    const filteredTodoList = todoState.todoList.filter((todo) => 
         todo.title.toLowerCase().includes(queryString.toLowerCase())
-     );
-
-    const indexOfFirstTodo = (currentPage - 1) * itemsPerPage;
-    const indexOfLastTodo = indexOfFirstTodo + itemsPerPage;
-    const currentTodos = filteredTodoList.slice(indexOfFirstTodo, indexOfLastTodo);
-
-    const totalPages = Math.ceil(filteredTodoList.length / itemsPerPage);
-
-    useEffect(() => {
-      if (totalPages > 0) {
-        if (isNaN(currentPage) || currentPage < 1 ||currentPage > totalPages) {
-            navigate("/"); 
-        } }
-    }, [currentPage, totalPages, navigate]);
-
-    const handlePreviousPage = () => {
-        if (currentPage > 1) {
-          setSearchParams({ page: (currentPage - 1).toString() });
-        }
-    };
-
-    const handleNextPage = () => {
-       if (currentPage < totalPages) {
-          setSearchParams({ page: (currentPage + 1).toString() });
-        }
-    };
+    );
 
     return (
         <div>
@@ -67,30 +38,7 @@ function TodosPage({
                 queryString={queryString}
                 setQueryString={setQueryString}
             />
-            <div
-                className="paginationControls"
-                style={{
-                display: "flex",
-                justifyContent: "center",
-                alignItems: "center",
-                gap: "1rem",
-                marginTop: "1rem"
-                }}
-            >
-                <button onClick={handlePreviousPage} disabled={currentPage === 1}>
-                  Previous
-                </button>
-
-                <span>
-                  Page {currentPage} of {totalPages || 1}
-                </span>
-
-                <button
-                  onClick={handleNextPage}
-                  disabled={currentPage === totalPages || totalPages === 0}
-                > Next
-                </button>
-            </div>
+            
         </div>
     )
 }
